Actualizar la versión de Java a la 21, refactorizar los repositorios para usar el tipo Long para los IDs  y agregar BaseRepository para funcionalidad de soft delete comun entre los repository.

+++ src/main/java/com/example/Muebleria/Controlador/CompraControlador.java
@@ -22,7 +22,7 @@ public class CompraControlador {
         return  servi.ObtenerTodo();
     }
     @GetMapping("/buscarid/{id}")
-    public Optional<CompraDto> Buscar(@PathVariable("id") Integer id){
+    public Optional<CompraDto> Buscar(@PathVariable("id") Long id){
         return  servi.Buscar(id);
     }
